compiler-attributes: rename header to match pattern
All other files in libnvme use a dash. Rename the compiler attributes
header accordingly.
