+++ libnvme/src/nvme/ioctl.c
@@ -23,7 +23,7 @@
 #include <libnvme.h>
 
 #include "private.h"
-#include "compiler_attributes.h"
+#include "compiler-attributes.h"
 
 static int nvme_verify_chr(struct libnvme_transport_handle *hdl)
 {
