+++ libnvme/src/nvme/log.c
@@ -20,7 +20,7 @@
 #include "cleanup.h"
 #define LOG_FUNCNAME 1
 #include "private.h"
-#include "compiler_attributes.h"
+#include "compiler-attributes.h"
 
 #ifndef LOG_CLOCK
 #define LOG_CLOCK CLOCK_MONOTONIC
