+++ libnvme/src/nvme/util.c
@@ -30,7 +30,7 @@
 #include "cleanup.h"
 #include "private.h"
 #include "util.h"
-#include "compiler_attributes.h"
+#include "compiler-attributes.h"
 
 /* The bionic libc implementation doesn't define LINE_MAX */
 #ifndef LINE_MAX
